Refactor navigation in CreateExercise screen
This commit refactors the exit functionality in the `CreateExercise` screen.

The `onExitOptionSelected` function in `CreateExerciseViewModel` and `CreateExerciseInteractionListener` has been replaced with two more specific functions: `onExitWithoutSavingClicked` and `onCancelClicked`.

The `CloseScreen` effect has been removed from `CreateExerciseEffect` as it's no longer needed. The navigation back is now handled directly in the `CreateExerciseScreen` when the `NavigateToAllExercises` effect is received.